Refactor engine.hpp
The engine has been simplified. Operators are defined in terms of
shared_ptr<Context>, instead of on Value.

This removed the need for custom hash functions that were needed when
traversing the graph. It also improved the readability of the code. The core of
engine.hpp is visible near the top of the file, and is clear and concise.

The `Value` struct is now a container for the shared pointers and defines the
friendly, pointerless API.